Resolves issues with 40 files ig
