--- src/main/java/org/skriptlang/skript/bukkit/entity/player/elements/events/EvtPlayerAnimation.java
@@ -1,48 +1,0 @@
-package org.skriptlang.skript.bukkit.entity.player.elements.events;
-
-import ch.njol.skript.lang.Literal;
-import ch.njol.skript.lang.SkriptEvent;
-import ch.njol.skript.lang.SkriptParser.ParseResult;
-import ch.njol.skript.lang.SyntaxStringBuilder;
-import org.bukkit.event.Event;
-import org.bukkit.event.player.PlayerAnimationEvent;
-import org.jetbrains.annotations.Nullable;
-import org.skriptlang.skript.bukkit.registration.BukkitSyntaxInfos;
-import org.skriptlang.skript.registration.SyntaxRegistry;
-
-public class EvtPlayerAnimation extends SkriptEvent {
-
-	public static void register(SyntaxRegistry syntaxRegistry) {
-		syntaxRegistry.register(BukkitSyntaxInfos.Event.KEY, BukkitSyntaxInfos.Event.builder(EvtPlayerAnimation.class, "Player Arm Swing")
-			.supplier(EvtPlayerAnimation::new)
-			.addEvent(PlayerAnimationEvent.class)
-			.addPatterns("[player] arm swing")
-			.addDescription("""
-				Called when a player swings their arm.
-				""")
-			.addExample("""
-				on player arm swing:
-				    send "You swung your arm!" to player
-				""")
-			.addSince("2.5.1")
-			.build());
-	}
-
-	@Override
-	public boolean init(Literal<?>[] args, int matchedPattern, ParseResult parseResult) {
-		return true;
-	}
-
-	@Override
-	public boolean check(Event event) {
-		return true;
-	}
-
-	@Override
-	public String toString(@Nullable Event event, boolean debug) {
-		return new SyntaxStringBuilder(event, debug)
-			.append("player arm swing")
-			.toString();
-	}
-
-}
