--- src/main/java/org/skriptlang/skript/bukkit/entity/player/elements/events/EvtPlayerTrade.java
@@ -1,55 +1,0 @@
-package org.skriptlang.skript.bukkit.entity.player.elements.events;
-
-import ch.njol.skript.lang.Literal;
-import ch.njol.skript.lang.SkriptEvent;
-import ch.njol.skript.lang.SkriptParser.ParseResult;
-import ch.njol.skript.lang.SyntaxStringBuilder;
-import io.papermc.paper.event.player.PlayerTradeEvent;
-import org.bukkit.entity.AbstractVillager;
-import org.bukkit.event.Event;
-import org.jetbrains.annotations.Nullable;
-import org.skriptlang.skript.bukkit.lang.eventvalue.EventValue;
-import org.skriptlang.skript.bukkit.lang.eventvalue.EventValueRegistry;
-import org.skriptlang.skript.bukkit.registration.BukkitSyntaxInfos;
-import org.skriptlang.skript.registration.SyntaxRegistry;
-
-public class EvtPlayerTrade extends SkriptEvent {
-
-	public static void register(SyntaxRegistry syntaxRegistry, EventValueRegistry eventValueRegistry) {
-		syntaxRegistry.register(BukkitSyntaxInfos.Event.KEY, BukkitSyntaxInfos.Event.builder(EvtPlayerTrade.class, "Player Trade")
-			.supplier(EvtPlayerTrade::new)
-			.addEvent(PlayerTradeEvent.class)
-			.addPatterns("[player] trad(e|ing)")
-			.addDescription("""
-				Called when a player trades with a villager or wandering trader.
-				""")
-			.addExample("""
-				on player trading:
-				    send "Did you get a good deal?" to player
-				""")
-			.addSince("2.7")
-			.build());
-
-		eventValueRegistry.register(EventValue.builder(PlayerTradeEvent.class, AbstractVillager.class)
-			.getter(PlayerTradeEvent::getVillager)
-			.build());
-	}
-
-	@Override
-	public boolean init(Literal<?>[] args, int matchedPattern, ParseResult parseResult) {
-		return true;
-	}
-
-	@Override
-	public boolean check(Event event) {
-		return true;
-	}
-
-	@Override
-	public String toString(@Nullable Event event, boolean debug) {
-		return new SyntaxStringBuilder(event, debug)
-			.append("player trade")
-			.toString();
-	}
-
-}
